Add Memory<byte> support for ACRCloudRecognizer. Split ACRCloudRecognizer.cs
